Add random component to domain/bucket name in test
This fixes bucket name collision in StageCreate test.

We need a random component in the domain name because it is used in bucket name and bucket names must be globally unique.

This uses a custom random number snippet. We can't use uuid because it would make the bucket name too long. Bucket name max length is 63 and uuid would result in a bucket name with ~70 characters.