Closes #580 - Add package.json based plugin discovery
Plugins are now discovered with the help of the projects package.json file. This makes it
possible to simply run 'npm install <pluginname> --save' to use plugins. No need to add the
plugin name to the plugins array in the s-project.json file (although this is still supported for backwards
compatibility).