Enable the use of SessionToken when it is available.
Add AWS session token usage and handling (for NodeJs or Python). This is important to enable the use of temporary credentials, particularly for cases of credential federation (e.g. SAML) where end users are unable to establish permanent IAM credentials.

Credentials are obtained as follows
1. Will use ~/.aws/credentials defined credentials if SERVERLESS_ADMIN_AWS_PROFILE is set
2. Checks for environmentally defined credentials with Serverless or standard AWS prefixes
3. Reads any duplicated credentials in admin.env.

If SERVERLESS_ADMIN_AWS_PROFILE exists but obtaining the credentials from ~/.aws/credentials fails, it falls back on using the environment or duplicated credentials. (as is the current logic, to avoid breaking existing behavior.)

The logic for obtaining credentials and preparing AWS SDK config has been consolidated. As appropriate, where defaults are inappropriate, customization of the base default is necessary/implemented.

Also adding some comments from gitter to the contribution notes. (I'd be happy to break that into a separate PR if desired)