Update README.md
Hello all,

I've been working on serverless-offline for a month now (:birthday:). It supports Velocity, responseParameters, serverless-babel-runtime, etc.. 
It's not an alternative to serverless-serve anymore (incompatible with 0.5), it's much more, so I wanted to change this line.

Nice job one 0.5, I like what you did a lot. When Google Function is out of alpha I'll be trying it so tell me if you need any help on supporting it.
Thanks,