Falsy event has no name attribute bugfix
Because it is falsy, event cannot have the attribute "name".  As such, a misleading undefined error is thrown rather than the more instructive error that the event couldn't be found for the region.  This leaves the actual identification of which event couldn't be found up to a good guess, unfortunately.