Add merging of generated resources to the aws.Resources section (in service)
This way the plugin simply adds the Resources to the CloudFormation template in the
compileEvents lifecycle. The deploy plugin will use this resources to update the stack later on.
This removes the dependency between the awsCompileS3Events plugin and the deploy plugin
as the awsCompileS3Events plugin simply adds the resources to the section and forgets about it.