Add subdirectory for tmpDirs
Add a subdirectory so that all directories which are generated for the tests are stored there.
Furthermore old tests are updated so that they use the tmpDir name generated by the test
utils functionality. This makes working with Docker way better as the tmpDirs are not created
It the root directory.