Add `options` to SDK, Object.assign(naming), merge
Make options available to the SDK.  This allows the options to be a source of configuration for the region and stage.  This allows them to override the provider (or be used in lieu of provider stage and region settings) in order to be used in names.
Add methods getRegion and getStage to SDK which centralizes knowledge of obtaining the correct values.  Use these instead of spreading correct source throughout codebase.
Add tests for new methods ^
Provide the naming methods to the SDK by `Object.assign()`ing them to `this.naming` (a new attribute that is an empty object except for a reference back to the SDK).  This sequesters the large number of naming methods while providing this.sdk.* access.  Maybe more importantly, it is consistent with the integration of logic in `lib/monitorStack.js` and `lib/validate`.