update sample profile section to use iamRoleARN
I'm running this:

     OS:                 darwin
     Node Version:       7.1.0
     Serverless Version: 1.1.0

and I can't get `serverless` to assign a role for my function (given a role ARN) *unless* I use the `iamRoleARN` yaml key, as shown in this patch.   

Using the `iamRoleARN` works.  I discovered it in someone's sample `serverless.yaml`, in an unrelated forum thread post.

I can't seem to find where in the code this value is consumed though... so please confirm this proposed patch is correct!