Closes #4687
1. Add a "PromiseTracker" to the Variables class. It provides tracking
and reporting of pending variable resolutions during service
populations.
2. Decompose populateObject and populateProperty into smaller pieces
that are easier to understand.
3. Additionally, remove the immediate recursive resolution of variables,
recursing only at the Object or Property entry point (when changes have
been made). This avoids a problem where the resolved value, prior to
being replaced in the containing object or value refers to itself and
thereby is waiting on its own resolution before it will resolve itself.
4. Return the given value in warnIfNotFound so that it can chain.