Add short wait before getting ssm variables
The issue described in #4311 comes from the fact, that aws needs `provider.profile` and `provider.stage` to be resolved before attempting to resolve the ssm references.
As there is no way to separate ssm variables from others, a short wait (10 ms) was added, before attempting to fetch the parameters from aws.