Kubeless: warn about bad label values
Add a warning in the documentation about using values that can be interpolated into variable types that are not strings by the serverless framework. Values for labels that are not strings are not supported in kubeless/kubernetes and basically break the kubeless deployment with a `ReadString: Labels: expected ...` forever because the pods will fail to start. This is serious because it seems the only way to recover is to remove the `kubeless` namespace in kubernetes (and as a result all the related primitives) and then redeploy `kubeless`. This was tested on kubeless `v1.0.0-alpha2`.