Attempt to fix inline editior
In the 'Use an existing AWS Profile` section, the inline editor on serverless.com is broken by the sample code. While I can't be sure, it may be due to the unbalanced use of special character asterisk.

Issue: https://imgur.com/a/UgzpkJ4

The attemped fix is change the number of asterisks to an even number (15->16) to see if that helps. other options would be to escape them.