Fix #5664 - Rollback fails due to a timestamp parsing error
When running sls rollback -t 1546961025152 the timestamp was being parsed as a string inside a Date Object.
This would return an Invalid Date, therefore, failing to convert the timestamp into a valid date.

In here we are checking to see if the date is already a valid one, otherwise convert the timestamp to a number and try again.
We are checking first because most cases we don't want to convert to numbers.

Fix #5664

Ps: We could add another check, if the date is still invalid, throw an Error, but that's not the scope of the issue.

Signed-off-by: Luan <luan@luanmuniz.com.br>