Revert "Don't add newlines to echoed input"
This reverts commit e30e24222bd9cd584e127d8cd9ea85386c614857.

This breaks a couple of tests which are depending on:

    serverless.cli.consoleLog.lastCall.args[0]

being set.

Also, it introduces an inconsistency c.f. invokeLocalPython et al.

I think here, consistency and passing tests trumps the slight annoyance of each input charact appearing on a new line. At least, for the moment and until a better way is found.