Check if role is in resources
If the role used by websockets lambda is managed it still tries to add a policy. Resulting in this error:
TypeError: Cannot read property 'Properties' of undefined
	at AwsCompileWebsockets.compileApi (/usr/local/lib/node_modules/serverless/lib/plugins/aws/package/compile/events/websockets/lib/api.js:35:8)

I ran into this error because my company is starting to lock down iam permissions and I do not have the iam:CreateRole, so I need to use an existing role.