Add support for VPC Endpoint associations in AWS
Amazon API Gateway support associating your API with an interface VPC endpoint. This is supported in cloudformation via EndpointConfiguration. Associating your API with the VPCE creates a custom DNS alias which allows you to access your private gateway through the VPCE without the need of a Host header, which is necessary in certain situations, particularly if you want to make ajax calls to your private API from a browser.