feat: manual API Gateway deployment dependencies
Adds the ability to add manual `DependsOn` entries in the
`AWS::ApiGateway::Deployment' resource.

Sometimes you need to ensure that the deployment waits for
resources defined in the `resources` section to be created.

This adds a `provider.apiGateway.deployment.dependsOn`
property that allows API creators to define dependencies
for the `AWS::ApiGateway::Deployment`.