feat: Support resource extensions
The merge logic for extending resources with the existing framework
using `resources.Resources` has some surprising behaviours.

This introduces a more formalized definition for extending resources, where
`Metadata` and `Properties` will be shallowly merged (the values for existing
keys will be replaced instead of merged), the `DependsOn` list will be added
to, and the values for `CreationPolicy`, `DeletionPolicy`, `UpdatePolicy`, and
`UpdateReplacePolicy` will be replaced if they exist in the extension.

Extension of other resource attributes is not supported at this time.