CosmosDB documentation: Show how to set http out
If you follow the example, requests to http endpoints always returns "204 No Content", independently of the status you set in "context.res". This is because the function.json generated doesn't has the output binding:

   {
      "type": "http",
      "direction": "out",
      "name": "res"
    }

This only happens if you combine http and cosmosDB events. Only with http and without cosmosDB, it works fine (only setting the first event without the http out event). Maybe is a bug, because with cosmosdb event, http event works in a different way than without.

It is necessary to set the http out event explicitly to work. This way you can return 201, 400, or whatever you need (and with the resulting json in the body if you want).

This occurs with other endpoints like GET too.