Allow parameters to be passed in
This change allows parameters to be passed in leveraging plugins.  There are occasions when parameters offer a better option than directly compiling values into the cloudformation.  One example of this is passwords.   If a password were embedded into a cloudformation, anyone with read access could see the password.  Using a parameter for this field with NoEcho, will result in the parameter being masked.