Include layer details in logical id digest
Currently the lambda version logical id is created by detecting any
changes in either the code or the lambda config.

However, if the lambda uses a layer that has updated then there will be
no change to the config as it just contains a reference to the resource
name.

To fix this, the layer details are extracted which will include the
timestamped S3 layer zip name and therefore generate a new digest if the
layer zip file name changes.