fix(AWS Lambda): Ensure version hash is in all cases effective (#8310)
In some instances the version hash may change due to a change involving
CloudFormation instrisic functions, but upon resolving the functions, the
resulting lambda configuration is identical to what is currently
deployed. This causes the Lambda::Version to fail to publish since AWS
sees the lambdas as identical.

This commit embeds an environment variable onto each lambda function with
the value of the version config hash, ensuring that anytime the config
hash changes, the resulting lambda configuration will be changed,
allowing a Lambda::Version to be published.