refactor: Abstract all `.serverless/` directory references
We want to be able to package / deploy serverless stacks in parallel.
But before we can do that we need to track down and abstract all the
places that the `.serverless/` directory is used.

The `.serverless/` directory is used as a temporary working directory,
but that prevents parallel Serverless commands from being run at the
same time with out colliding. A unique temporary directory is needed for
each run. However, in order to maintain backwards compatibility the
framework will need to default to using the `.serverless` directory as
plugins may hook in at various points in the compilation, packaging,
deploying, etc. to read or write files in the `.serverless/` directory.

With this change, all uses of `.serverless` either point to the package
output directory (getPackageOutputDir), or to the temporary working
directory (getWorkingTempDir).