docs: Suggest `python3` template over `python2` (#9773)
python2 is [no longer a supported runtime](https://aws.amazon.com/blogs/compute/announcing-end-of-support-for-python-2-7-in-aws-lambda/) for AWS.

When I ran this example as it currently is in the docs it created a python2 project. Interestingly selecting `AWS - Python - Starter` from the CLI did create a python3 project, although the code seems to suggest this should be calling the `aws-python` template, but I probably missed a link there.