docs: Removed duplicated paragraph
The following paragraph was duplicated in many sections of the page:

> **Note:** The `stream` event will hook up your existing streams to a Lambda function. Serverless won't create a new stream for you.

I think this was because of contributors copy/pasting that along the way. When reading the page it felt really weird reading that over and over.

If there was a reason this paragraph was there in every section feel free to close.