Config schema: add `websocketsDescription`
I used it back with the v1 version, and now on v2 (at least) I got a _Configuration warning_ when defining that value:

```
Serverless: Configuration warning at 'provider': unrecognized property 'websocketsDescription'
```

The value is used in different places but is missing in the config schema.
I've also updated the doc about it.