feat: allow serverless configuration with .cjs and .mjs extensions
This commit adds support for .cjs and .mjs file extensions on the serverless
configuration file. If defined as `serverless.mjs`, the file will be sourced
using dynamic import instead of require.js.

This encompasses a minimal viable change to allow the user to use es-modules
in his codebase. While we don't check the user's package.json for the package
type yet, it should still allow using an es-module project by simply setting
the configuration extension to .cjs or .mjs explicitly.