Update serverless.js
The script uses the `triage` function provided by the `cli` library to determine the command that is being run. If the command is `serverless`, it runs the serverless script. If the command is `@serverless/compose`, it runs the `run-compose` function, which is wrapped in a promise. If the command is `serverless-tencent`, the script prints a message indicating that the command is no longer supported and exits with a non-zero status.

Notice that I added a break statement after each cases of switch statement, this ensures that the script doesn't "fall through" and execute the following cases if the current one is a match. Also, I used destructuring for variables that were used multiple times and I moved constant values to the top of the script and commented to explain their role in the script.

The Last section of the code adds additional functionality to handle custom commands and invalid commands. The script first checks if the input command is "my-custom-command", if so it will execute some custom code, otherwise it will check that the command is a valid command, if it is not then the script will return an error message and exit.

The code uses the `minimist` library to parse the input command line arguments and separates them into options and flags. This allows for more flexibility in handling different types of input. It also has a try-catch block to handle any error that might happen during the argument parsing process.

The script now checks for a custom command flag `debug` and execute different action depending on if the flag is present or not. Also, handling different options and flags depending on the command that was passed.