Update apigateway.md
errorMessage": "Expected `schema` to be an object or boolean", "errorType": "SchemaError", "offlineInfo": "If you believe this is an issue with serverless-offline please submit it, thanks. https://github.com/dherault/serverless-offline/issues", "stackTrace": [ "SchemaError: Expected `schema` to be an object or boolean", "at Validator.validate
I was getting above error due to below configuration of my yml file . due to serverless-offline update schemas is changed to schema but  document is not upto date. So i have changed schemas to schema inside the function.
provider:
    ...
    apiGateway:
      request:
        schemas:
          post-create-model:
            name: PostCreateModel
            schema: ${file(api_schema/post_add_schema.json)}
            description: "A Model validation for adding posts"

functions:
  create:
    handler: posts.create
    events:
      - http:
          path: posts/create
          method: post
          request:
            schemas:
              application/json: post-create-model